Adding custom modules imports
Check if view controller belongs to a custom Module and imports it if needed so that the project can compile. e.g.(AVPlayerViewController from AVKit needs the "import AVKit" line).

TODO: For cosmetic purposes, we need check for duplicated imports, but that requires a bigger refactor.
I would suggest that we could hold the output before printing it, for post processing string analysis (duplicated imports for instance). I'll write a issue to address that and further discussion.